Cache item labels to prevent exception due to missing db connection.
The cause of this issue is the asynchronous execution of a filter
depending on the computation of item labels.
This filter is executed on a thread with no associated active
unit of work and therefore unable to access the database
to retrieve information about item labels (i.e. rdfs:label).

Change-Id: Ia8033668c2f9b86eb3bf992f977c5c060a717f7a